+++ src/services/answer-refiner/tests/answer-refiner.test.ts
@@ -1,8 +1,8 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { AnswerRefiner } from "@/services/answer-refiner";
-import type { GeminiClient } from "../../gemini-client";
-import type { GoogleSheetsService } from "../../google-sheets";
+import type { GeminiClient } from "@/services/gemini-client";
+import type { GoogleSheetsService } from "@/services/google-sheets";
 
 // Mock dependencies
 const mockGeminiClient = {
